Improve WeatherForecast.TemperatureF testing and fix overflow bug
- Added comprehensive unit tests in `Archiver.Services.Tests/WeatherForecastTests.cs` covering standard range and edge cases.
- Identified and fixed an integer overflow bug in `Archiver.Shared/WeatherForecast.cs` by casting to `long` during intermediate calculation.
- Verified the fix with new regression tests.

+++ Archiver.Shared/WeatherForecast.cs
@@ -4,5 +4,5 @@ public readonly record struct WeatherForecast(DateOnly Date, int TemperatureC, s
 {
     private const int FahrenheitFreezingPoint = 32;
 
-    public int TemperatureF => FahrenheitFreezingPoint + (TemperatureC * 9 / 5);
+    public int TemperatureF => FahrenheitFreezingPoint + (int)((long)TemperatureC * 9 / 5);
 }
